Consistent base ability class
The admin plugin uses a namespaced `BaseAbility` class. Let's do this
in core as well for consistency.
